requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-dialogflow==2.41.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
